Add simple safety checks on changing player count
Checks to ensure that a maps player count never underflow, and that it never exceeds 40. 40 was chosen as the max as that is the largest possible raid size as far as I am aware, and it saves a bit on memory to check vs the absolute cap rather than the maps maximum, as we do not need to fetch the map itself.